1. fixed issue on update empty value / 0 / false will missing info to db. 2. add template bing/unbing to hostgroup api

+++ app/model/falcon_portal/strategy.go
@@ -30,7 +30,7 @@ type Strategy struct {
 	Note       string `json:"note" gorm:"column:note"`
 	RunBegin   string `json:"run_begin" gorm:"column:run_begin"`
 	RunEnd     string `json:"run_end" gorm:"column:run_end"`
-	TplId      uint   `json:"tpl_id" gorm:"column:tpl_id"`
+	TplId      int64  `json:"tpl_id" gorm:"column:tpl_id"`
 }
 
 func (this Strategy) TableName() string {
